Ensure empty blocks are parsed
Ultimately, the reason for this change is so that fontification runs
on a newly-created empty block (the fontification is responsible for
setting the cursor-sensor properties). Without this change, an empty
block that is indented, like

  \begin{align*}

  \end{align*}

was not being parsed, because skip-chars-forward positioned point
immediately before "\end", i.e. not in the first column, whereas the
math block regexes are anchored at the line-beginning.