Fix `xenops-src-parse-at-point` for Org 9.7
Fixes #73. The error seems to come from `org-element-context` containing
references to the killed temp buffer created by the `xenops-src-do-in-org-mode`
macro, and `org-babel-get-src-block-info` feeding those references into
`with-current-buffer`. Hence this fix just includes the call to
`org-babel-get-src-block-info` in the scrope of the `xenops-src-do-in-org-mode`
macro.

As I do not use xenops-mode myself, and thus do not know the expected behaviour
well, it is not very thoroughly tested.