Preserve window configuration in `xenops-do-src-in-org-mode`
As the current buffer gets narrowed in the macro, this can change the
`window-start` position. In particular, it scrolls the buffer such that the
current src block is at the top of the window. This side effect might be
undesirable (but feel free to disregard this if it is intended, or for any other
reason). Hence we prevent it with `save-window-excursion`.